+++ lib/jkf.rb
@@ -1,7 +1,7 @@
 require 'kconv'
-require 'jkf/version'
-require 'jkf/parser'
-require 'jkf/converter'
+require_relative 'jkf/version'
+require_relative 'jkf/parser'
+require_relative 'jkf/converter'
 
 # JSON Kifu Format
 module Jkf