+++ lib/jkf/parser.rb
@@ -6,8 +6,8 @@ module Jkf
   end
 end
 
-require 'jkf/parser/base'
-require 'jkf/parser/kifuable'
-require 'jkf/parser/kif'
-require 'jkf/parser/ki2'
-require 'jkf/parser/csa'
+require_relative 'parser/base'
+require_relative 'parser/kifuable'
+require_relative 'parser/kif'
+require_relative 'parser/ki2'
+require_relative 'parser/csa'