+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed link to json-kifu-format
 * Remove Inch CI link on readmes
 * Lint Ruby sources
+* Add benchmark script
 
 ## [0.5.0] - 2023-06-18